refactor(fetch): move web_fetch to its own capability dir
environments/local/ is organized by capability (search/, browser/,
filesystem/). web_fetch is URL retrieval, not search, so it gets its own
environments/local/fetch/ dir (FetchToolProvider) rather than living under
search/. Its activation is still gated on the firecrawl search provider, which
is config wiring handled by the orchestrator.
