+++ multimodal/agent-tars/core/src/environments/local/fetch/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './search-tool';
+export * from './fetch-tool';
